requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
googlemaps==4.10.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.0.7
